# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Culbridge Compliance API",
    "description": "Deterministic compliance engine for agricultural export compliance\n\n## Key Features\n- Deterministic rule evaluation (no probabilistic outcomes)\n- Full audit trail for regulatory compliance\n- Multi-corridor support (EU, USA, etc.)\n- Lab and document verification\n- HS Code validation\n\n## Error Handling\nAll endpoints return standardized error responses. See `EngineError` schema.",
    "version": "1.0.0",
    "contact": {
      "name": "Culbridge API Support",
      "email": "support@culbridge.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api.culbridge.com/v1",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/shipments": {
      "get": {
        "tags": ["Shipments"],
        "summary": "List all shipments",
        "description": "Returns paginated list of shipments. Filters by status, product, corridor.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by shipment status",
            "schema": { "type": "string", "enum": ["DRAFT", "PENDING_VALIDATION", "VALIDATING", "READY_TO_SUBMIT", "SUBMITTED", "UNDER_REVIEW", "APPROVED", "REJECTED", "SIGNED"] }
          },
          {
            "name": "product",
            "in": "query",
            "description": "Filter by product category",
            "schema": { "type": "string", "enum": ["cocoa", "sesame", "cashew", "ginger", "groundnuts", "seeds"] }
          },
          {
            "name": "corridor",
            "in": "query",
            "description": "Filter by destination corridor",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "List of shipments",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ShipmentListResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EngineError" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EngineError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Shipments"],
        "summary": "Create new shipment",
        "description": "Creates a new shipment in DRAFT status. Use /shipments/{id}/evaluate for compliance check.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ShipmentInput" },
              "example": {
                "product": "cocoa beans",
                "category": "cocoa",
                "hs_code": "1801.00.10",
                "origin_country": "NG",
                "destination_country": "DE",
                "batch_number": "CB-2026-001",
                "quantity_kg": 25000,
                "exporter_id": "EXP-001"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Shipment created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Shipment" }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EngineError" }
              }
            }
          }
        }
      }
    },
    "/shipments/{id}": {
      "get": {
        "tags": ["Shipments"],
        "summary": "Get shipment by ID",
        "description": "Returns full shipment details including all module outputs",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "description": "Shipment ID", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Shipment details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ShipmentDetail" }
              }
            }
          },
          "404": {
            "description": "Shipment not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EngineError" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Shipments"],
        "summary": "Update shipment",
        "description": "Update shipment fields. Only allowed for DRAFT status.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ShipmentUpdateInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Shipment updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Shipment" }
              }
            }
          },
          "400": {
            "description": "Update not allowed (not DRAFT)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EngineError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Shipments"],
        "summary": "Delete shipment",
        "description": "Delete shipment. Only allowed for DRAFT status.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Shipment deleted" },
          "400": {
            "description": "Cannot delete (not DRAFT)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EngineError" }
              }
            }
          }
        }
      }
    },
    "/shipments/{id}/evaluate": {
      "post": {
        "tags": ["Compliance"],
        "summary": "Evaluate shipment compliance",
        "description": "Runs deterministic compliance evaluation. Returns BLOCKERS, WARNINGS, and audit log.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Evaluation result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EvaluationResult" },
                "example": {
                  "shipment_id": "CB-001",
                  "status": "SAFE",
                  "score": 95,
                  "blockers": [],
                  "warnings": [
                    {
                      "code": "LOW_CONFIDENCE_HS_CODE",
                      "message": "HS Code 1801.00 has 65% confidence",
                      "severity": "WARNING"
                    }
                  ],
                  "audit_log": [
                    {
                      "rule_id": "EU_COCOA_AFLATOXIN_001_v2",
                      "input": { "value": 3.5, "unit": "μg/kg" },
                      "result": "PASS",
                      "timestamp": "2026-03-28T10:23:00Z"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/shipments/{id}/submit": {
      "post": {
        "tags": ["Shipments"],
        "summary": "Submit shipment for processing",
        "description": "Submits shipment to NSW ESB. Requires READY_TO_SUBMIT status.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Submission result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SubmissionResult" }
              }
            }
          },
          "400": {
            "description": "Not ready for submission",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EngineError" }
              }
            }
          }
        }
      }
    },
...
    "/documents/upload": {
      "post": {
        "tags": ["Documents"],
        "summary": "Upload document",
        "description": "Upload document (PDF, JPG, PNG, DOCX). Max 20MB.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file", "shipment_id", "document_type"],
                "properties": {
                  "file": { "type": "string", "format": "binary" },
                  "shipment_id": { "type": "string" },
                  "document_type": { "type": "string", "enum": ["COO", "phytosanitary", "health_cert", "lab_report", "packing_list", "invoice"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document uploaded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DocumentUploadResult" }
              }
            }
          },
          "413": {
            "description": "File too large",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EngineError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Shipment": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "description": "Unique shipment identifier (e.g., CB-001)" },
          "status": { 
            "type": "string", 
            "enum": ["DRAFT", "PENDING_VALIDATION", "VALIDATING", "READY_TO_SUBMIT", "SUBMITTED", "UNDER_REVIEW", "APPROVED", "REJECTED", "SIGNED"] 
          },
          "product": { "type": "string", "description": "Product name" },
          "category": { "type": "string", "enum": ["cocoa", "sesame", "cashew", "ginger", "groundnuts", "seeds"] },
          "hs_code": { "type": "string", "description": "Harmonized System code (6-10 digits)" },
          "origin_country": { "type": "string", "description": "ISO 2-letter country code" },
          "destination_country": { "type": "string", "description": "Destination corridor/country" },
          "batch_number": { "type": "string" },
          "quantity_kg": { "type": "number" },
          "exporter_id": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "ShipmentDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/Shipment" },
          {
            "type": "object",
            "properties": {
              "hs_code_validation": { "$ref": "#/components/schemas/HSCodeValidation" },
              "documents": { "type": "array", "items": { "$ref": "#/components/schemas/Document" } },
              "compliance": { "$ref": "#/components/schemas/ComplianceStatus" },
              "fees": { "$ref": "#/components/schemas/FeeCalculation" },
              "signature": { "$ref": "#/components/schemas/SignatureStatus" },
              "audit_log": { "type": "array", "items": { "$ref": "#/components/schemas/AuditLogEntry" } }
            }
          }
        ]
      },
      "ShipmentInput": {
        "type": "object",
        "required": ["product", "category", "hs_code", "origin_country", "destination_country", "exporter_id"],
        "properties": {
          "product": { "type": "string", "description": "Product name" },
          "category": { "type": "string", "enum": ["cocoa", "sesame", "cashew", "ginger", "groundnuts", "seeds"] },
          "hs_code": { "type": "string", "pattern": "^[0-9]{6,10}$", "description": "6-10 digit HS code" },
          "origin_country": { "type": "string", "minLength": 2, "maxLength": 2 },
          "destination_country": { "type": "string", "description": "Destination country code" },
          "batch_number": { "type": "string" },
          "quantity_kg": { "type": "number", "minimum": 0 },
          "exporter_id": { "type": "string" },
          "agent_id": { "type": "string", "description": "Optional agent" }
        }
      },
      "ShipmentUpdateInput": {
        "type": "object",
        "properties": {
          "product": { "type": "string" },
          "hs_code": { "type": "string" },
          "destination_country": { "type": "string" },
          "quantity_kg": { "type": "number" }
        }
      },
      "ShipmentListResponse": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Shipment" } },
          "pagination": { "$ref": "#/components/schemas/Pagination" }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "pages": { "type": "integer" }
        }
      },
      "EvaluationResult": {
        "type": "object",
        "properties": {
          "shipment_id": { "type": "string" },
          "status": { "type": "string", "enum": ["BLOCKED", "HIGH_RISK", "SAFE"] },
          "score": { "type": "integer", "minimum": 0, "maximum": 100 },
          "blockers": { "type": "array", "items": { "$ref": "#/components/schemas/EngineError" }, "description": "Critical errors that block submission" },
          "warnings": { "type": "array", "items": { "$ref": "#/components/schemas/EngineError" }, "description": "Non-critical issues" },
          "audit_log": { "type": "array", "items": { "$ref": "#/components/schemas/AuditLogEntry" } },
          "threshold_adjustments": { "type": "array", "items": { "type": "object" }, "description": "Dynamic threshold changes applied" },
          "rasff_alerts": { "type": "array", "items": { "type": "object" }, "description": "Matched RASFF alerts" }
        }
      },
      "ComplianceStatus": {
        "type": "object",
        "properties": {
          "eudr_compliant": { "type": "boolean" },
          "rasff_flags": { "type": "array", "items": { "type": "object" } },
          "lab_results": { "type": "array", "items": { "$ref": "#/components/schemas/LabResult" } },
          "violations": { "type": "array", "items": { "type": "object" } }
        }
      },
      "HSCodeValidation": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "confidence": { "type": "number", "minimum": 0, "maximum": 1 },
          "alternatives": { "type": "array", "items": { "type": "string" } },
          "validated": { "type": "boolean" }
        }
      },
...
      "AuditLogEntry": {
        "type": "object",
        "description": "Immutable audit log entry for regulatory compliance",
        "properties": {
          "event_id": { "type": "string", "description": "Unique event identifier" },
          "rule_id": { "type": "string", "description": "Rule that triggered this event" },
          "input": {
            "type": "object",
            "properties": {
              "value": { "type": "number" },
              "unit": { "type": "string" },
              "lab_id": { "type": "string" },
              "document_id": { "type": "string" }
            }
          },
          "result": { "type": "string", "enum": ["PASS", "FAIL", "BLOCKER", "WARNING"] },
          "timestamp": { "type": "string", "format": "date-time", "description": "ISO 8601 timestamp" },
          "actor": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "name": { "type": "string" },
              "role": { "type": "string" }
            }
          },
          "context": { "type": "object", "description": "Additional context (shipment_id, version, etc.)" },
          "hash": { "type": "string", "description": "SHA-256 hash for integrity verification" }
        }
      },
      "FeeCalculation": {
        "type": "object",
        "properties": {
          "shipment_id": { "type": "string" },
          "total": { "type": "number", "description": "Total fee in Naira" },
          "breakdown": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "certificate_id": { "type": "string" },
                "agency_name": { "type": "string" },
                "fee_naira": { "type": "number" },
                "processing_days": { "type": "integer" }
              }
            }
          },
          "estimated_processing_days": { "type": "object", "properties": { "min": { "type": "integer" }, "max": { "type": "integer" } } },
          "critical_path_days": { "type": "integer" }
        }
      },
      "SubmissionResult": {
        "type": "object",
        "properties": {
          "shipment_id": { "type": "string" },
          "sgd_number": { "type": "string", "description": "NSW-generated SGD number" },
          "submission_status": { "type": "string" },
          "submitted_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "LabResult": {
        "type": "object",
        "properties": {
          "lab_id": { "type": "string" },
          "test_type": { "type": "string" },
          "result": { "type": "number" },
          "unit": { "type": "string" },
          "limit": { "type": "number" },
          "status": { "type": "string", "enum": ["PASS", "FAIL"] },
          "test_date": { "type": "string", "format": "date" }
        }
      },
      "Document": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "shipment_id": { "type": "string" },
          "document_type": { "type": "string", "enum": ["COO", "phytosanitary", "health_cert", "lab_report", "packing_list", "invoice"] },
          "file_name": { "type": "string" },
          "file_hash": { "type": "string", "description": "SHA-256 hash" },
          "status": { "type": "string", "enum": ["pending", "verified", "rejected"] },
          "uploaded_at": { "type": "string", "format": "date-time" }
        }
      },
      "DocumentUploadResult": {
        "type": "object",
        "properties": {
          "document_id": { "type": "string" },
          "storage_path": { "type": "string" },
          "hash": { "type": "string" },
          "uploaded_at": { "type": "string", "format": "date-time" }
        }
      },
      "SignatureStatus": {
        "type": "object",
        "properties": {
          "signature_id": { "type": "string" },
          "signature_type": { "type": "string" },
          "signer_identity": { "type": "string" },
          "signing_time": { "type": "string", "format": "date-time" },
          "verified": { "type": "boolean" }
        }
      },
...
      "EngineError": {
        "type": "object",
        "description": "Standardized error object for all engine errors",
        "properties": {
          "name": { "type": "string", "description": "Error class name" },
          "severity": { "type": "string", "enum": ["BLOCKER", "WARNING", "INFO"], "description": "Error severity level" },
          "code": { "type": "string", "description": "Unique error code (e.g., API_TIMEOUT, MISSING_DOCUMENT)" },
          "message": { "type": "string", "description": "Human-readable message" },
          "context": { "type": "object", "description": "Additional context (shipmentId, labId, etc.)" },
          "timestamp": { "type": "string", "format": "date-time" },
          "stack": { "type": "string", "description": "Error stack trace (dev only)" }
        },
        "required": ["severity", "code", "message", "timestamp"]
      },
...
    }
  }
}